add tests for the -r option
Getting this to work was tricky. Essentially where we would usually simply mock the IO::Socket::INET->new method, we can't, because that code is in a different program. Test::Script runs the script in a different perl interpreter, so we can't influence anything in there. I had to bend over backwards to get this to work, but it's pretty ok. Read the code and the comments to see how please.

In theory this could be used to add extra test, like checking if the correct arguments were passed to IO::Socket. Especially the port would be relevant. Getting that out to test it would be hard though. A possibility could be to pass the expected port in and die if it's wrong, failing the tests in a pretty hard way.

It could look like this:

    "-MSocketOverride=$state,$port"

However, this is indeed insane.